refactor: move and rename adapter
